+++ app/stream/[slug]/GroupStreamView.tsx
@@ -5,7 +5,6 @@ import { NoProducersView } from "./NoProducersView"
 
 export function GroupStreamView(producerCount: number, group: Group) {
     return <div className={"m-10"}>
-        {producerCount > 0 ? <StreamView producers={group.producers} streamName={group.name}/> :
-            <NoProducersView group={group}/>}
+        {producerCount > 0 ? <StreamView producers={group.producers} streamName={group.name}/> : <NoProducersView group={group}/>}
     </div>;
 }